+++ database/migrations/2019_11_25_160614_create_m_alamat_table.php
@@ -14,7 +14,7 @@ class CreateMAlamatTable extends Migration {
 	{
 		Schema::create('m_alamat', function(Blueprint $table)
 		{
-			$table->integer('id', true);
+			$table->bigint('id', true);
 			$table->integer('pid')->nullable();
 			$table->string('nama')->nullable();
 			$table->string('jenis')->nullable();
